script/requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
pyautogui>=0.9.54
pygetwindow>=0.0.9
pydirectinput>=1.0.4
pyperclip>=1.9.0
pyscreeze>=1.0.1
Pillow>=11.0.0
opencv-python>=4.10.0.84
mss>=10.0.0
rapidocr-onnxruntime==1.2.3
